fix firefox automation and adress feedback
