+++ packages/app/src/runner/automation/AutomationElement.vue
@@ -1,6 +1,6 @@
 <template>
   <div
-    :id="automationElementId"
+    :id="runnerUiStore.automationElementId"
     style="display: none"
   >
     {{ runnerUiStore.randomString }}
@@ -10,6 +10,5 @@
 <script lang="ts" setup>
 import { useRunnerUiStore } from '../../store'
 
-const automationElementId = window.__CYPRESS_CONFIG__.namespace
 const runnerUiStore = useRunnerUiStore()
 </script>
