+++ packages/app/src/store/runner-ui-store.ts
@@ -1,3 +1,4 @@
+import type { AutomationElementId } from '@packages/types'
 import { defineStore } from 'pinia'
 import { runnerConstants } from '../runner/runner-constants'
 
@@ -26,6 +27,7 @@ export interface RunnerUiState {
   reporterWidth: number
   automationStatus: AutomationStatus
   randomString: string
+  automationElementId: AutomationElementId
 }
 
 export const useRunnerUiStore = defineStore({
@@ -40,6 +42,7 @@ export const useRunnerUiStore = defineStore({
       reporterWidth: runnerConstants.defaultReporterWidth,
       automationStatus: automation.CONNECTING,
       randomString: `${Math.random()}`,
+      automationElementId: `${window.__CYPRESS_CONFIG__.namespace}-string`,
     }
   },
